UnTypeHandleを使用して型識別なしでAssetBaseを取得可能にしました。

+++ GameEngine/include/Assets/Managers/AssetRegistry.h
@@ -63,7 +63,7 @@ namespace REngine
 			if (m_freeIndexList.empty())
 			{
 				// インデックスを末尾に
-				index = m_slots.size();
+				index = static_cast<uint32_t>(m_slots.size());
 
 				// 空のスロットを生成
 				m_slots.push_back({});
